+++ src/api/v1/endpoints/login.py
@@ -179,8 +179,8 @@ async def refresh_token(
         raise HTTPException(status_code=404, detail="Incorrect token")
 
 
-@router.post("/token", response_model=TokenRead)
-async def login_access_token(
+@router.post("/token", response_model=Token)
+async def oauth2_login(
     form_data: OAuth2PasswordRequestForm = Depends(),
     redis_client: Redis = Depends(get_redis_client),
 ) -> Any:
@@ -189,11 +189,19 @@ async def login_access_token(
     """
     user = await crud.user.authenticate(email=form_data.username, password=form_data.password)
     if not user:
-        raise HTTPException(status_code=400, detail="Incorrect email or password")
+        raise HTTPException(status_code=400, detail="Email or Password incorrect")
     elif not user.is_active:
-        raise HTTPException(status_code=400, detail="Inactive user")
+        raise HTTPException(status_code=400, detail="User is inactive")
     access_token_expires = timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
+    refresh_token_expires = timedelta(minutes=settings.REFRESH_TOKEN_EXPIRE_MINUTES)
     access_token = security.create_access_token(user.id, expires_delta=access_token_expires)
+    refresh_token = security.create_refresh_token(user.id, expires_delta=refresh_token_expires)
+    data = Token(
+        access_token=access_token,
+        token_type="bearer",
+        refresh_token=refresh_token,
+        user=user,
+    )
     valid_access_tokens = await get_valid_tokens(redis_client, user.id, TokenType.ACCESS)
     if valid_access_tokens:
         await add_token_to_redis(
@@ -203,7 +211,13 @@ async def login_access_token(
             TokenType.ACCESS,
             settings.ACCESS_TOKEN_EXPIRE_MINUTES,
         )
-    return {
-        "access_token": access_token,
-        "token_type": "bearer",
-    }
+    valid_refresh_tokens = await get_valid_tokens(redis_client, user.id, TokenType.REFRESH)
+    if valid_refresh_tokens:
+        await add_token_to_redis(
+            redis_client,
+            user,
+            refresh_token,
+            TokenType.REFRESH,
+            settings.REFRESH_TOKEN_EXPIRE_MINUTES,
+        )
+    return data